requirements.txt
 
		 tkinter
		 messagebox
		 PIL
		 cv2
		 argparse
	    matplotlib.pyplot
	 	 numpy
		 time
		 os
		 tensorflow
		 fpdf